Refactored blaster stuff into blaster file (this annoyed me)
