+++ src/main.rs
@@ -32,8 +32,8 @@ fn main() {
         .add_systems(Update, gameplay::player::flight.after(gameplay::player::move_player))
         .add_systems(Update, gameplay::player::animate_player.after(gameplay::player::move_player))
         .add_systems(Update, ui::camera::move_camera.after(gameplay::player::move_player))
-        .add_systems(Update, gameplay::player::update_blaster_aim)
-        .add_systems(Update, gameplay::player::shoot_blaster.after(gameplay::player::update_blaster_aim))
+        .add_systems(Update, gameplay::blaster::update_blaster_aim)
+        .add_systems(Update, gameplay::blaster::shoot_blaster.after(gameplay::blaster::update_blaster_aim))
         .add_systems(Update, engine::particles::Particle::move_and_handle_collisions.after(gameplay::player::flight))
         .add_systems(Update, gameplay::enemy::enemy_gravity)
         .add_systems(Update, gameplay::enemy::track_player)
